+++ client-app/src/mobile/treegrids/TreeGridPageModel.js
@@ -2,6 +2,7 @@ import {HoistModel, managed, XH} from '@xh/hoist/core';
 import {GridModel} from '@xh/hoist/cmp/grid';
 import {millionsRenderer, numberRenderer} from '@xh/hoist/format';
 import {GroupingChooserModel} from '@xh/hoist/mobile/cmp/grouping';
+import {isEmpty} from 'lodash';
 
 export class TreeGridPageModel extends HoistModel {
 
@@ -21,8 +22,18 @@ export class TreeGridPageModel extends HoistModel {
         },
         colChooserModel: true,
         sortBy: 'pnl|desc|abs',
-        onRowClicked: (e) => {
-            const id = encodeURIComponent(e.data.raw.id);
+        onRowClicked: ({data: record}) => {
+            // Allow single tap on parent row to trigger expand/collapse (default on mobile)
+            // without navigating the user away to the detail page.
+            if (!isEmpty(record.children)) return;
+
+            const id = encodeURIComponent(record.id);
+            XH.appendRoute('treeGridDetail', {id});
+        },
+        // Used here (and recommended) as an alternate gesture to support drilldown on parent rows
+        // when `GridModel.clicksToExpand: 1` (the default). Fires on long-press / tap-and-hold.
+        onCellContextMenu: ({data: record}) => {
+            const id = encodeURIComponent(record.id);
             XH.appendRoute('treeGridDetail', {id});
         },
         columns: [